+++ client/App.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Carousel from './components/Carousel.jsx';
+
+ReactDOM.render(<Carousel />, document.getElementById('app'));
