+++ package.json
@@ -10,7 +10,13 @@
   "scripts": {
     "react-dev": "webpack -d --watch",
     "server-dev": "nodemon server/index.js",
-    "seed-db": "mongoimport --db restaurants --collection Nearby --file nearby-data.json"
+    "seed-db": "mongoimport --db restaurants --collection Nearby --file nearby-data.json",
+    "test": "jest"
+  },
+  "jest": {
+    "setupFilesAfterEnv": [
+      "<rootDir>client/setupTests.js"
+    ]
   },
   "dependencies": {
     "express": "^4.17.1",
@@ -21,12 +27,15 @@
     "react-icons": "^3.7.0"
   },
   "devDependencies": {
-    "nodemon": "^1.19.1",
     "@babel/core": "^7.5.0",
     "@babel/preset-env": "^7.5.2",
     "@babel/preset-react": "^7.0.0",
     "babel-loader": "^8.0.6",
+    "enzyme": "^3.10.0",
+    "enzyme-adapter-react-16": "^1.14.0",
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
+    "jest": "^24.8.0",
+    "nodemon": "^1.19.1",
     "webpack": "^4.35.3",
     "webpack-cli": "^3.3.5"
   }
