+++ webpack.config.js
@@ -3,7 +3,7 @@ var SRC_DIR = path.join(__dirname, '/client');
 var DIST_DIR = path.join(__dirname, '/public');
 
 module.exports = {
-  entry: `${SRC_DIR}/Carousel.jsx`,
+  entry: `${SRC_DIR}/App.jsx`,
   output: {
     filename: 'bundle.js',
     path: DIST_DIR
